Extract executable optimizer tests
